requirements.txt
certifi==2022.5.18.1
Pillow==9.0.1
wincertstore==0.2
